requirements.txt
streamlit
pandas
numpy
matplotlib
sklearn
pillow
opencv-python-headless